Updated wizard and edit sceens
Made some updates to the wizard. The dynamic classes now pull from the
schedule table. The edit screen changes the values in the schedule table
and updates the label text. Also added an unassigned room that is
important to get the edit screen working. And created several stored
procs to update the schedule table and get teacher information.