Ability to add in same class times (only d & d)
Added in the ability to drag and drop classes at the same time. It
should allow any number of classes. Some iffy things with weird cases,
but for the most part it works. Need to add in to pull up edit screen
and write to database. Try to break this if you can and record it.