More edit functionality
The edit form can now add a class to the schedule if there isn't one in
that spot already. Also added in more usability with drag and drop and
writing to database. Technically, we can create a whole schedule now.
Not done yet. *Needs the new stored proc to work.