Schedule Table Update
Update to the schedule table caused many stored procs to be changed.
Will need to rerun all scripts if using your local database. Also, added
in the ability to remove a semester and year from the schedule table if
needed to create a new one. **Should run test data on a previous
semester and year to test the reports stuff.